# backend/moni/openapi-schema.yml
openapi: 3.0.2
info:
  title: "Moni"
  version: "0.0.0"
  description: |
    Moni - A HTTP service monitoring solution OpenAPI schema.
    Checkout project on [Github](https://github.com/hibare/Moni).
  license:
    name: MIT
    url: https://github.com/hibare/Moni/blob/master/LICENSE
...
paths:
...
  /api/v1/jobs/{uuid}/:
    get:
      operationId: Job details
      description: Get a Job details
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jobs"
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "No such job"
      tags:
        - Jobs
    put:
      operationId: Update Job
      description: Update a Job
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Jobs"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Jobs"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Jobs"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jobs"
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "No such job"
      tags:
        - Jobs
    patch:
      operationId: Partial update Job
      description: Partial update a Job
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Jobs"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Jobs"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Jobs"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jobs"
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "No such job"
      tags:
        - Jobs
    delete:
      operationId: Delete Job
      description: Delete a Job
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      responses:
        "204":
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "No such job"
      tags:
        - Jobs
  /api/v1/jobs/{uuid}/pause/:
    post:
      operationId: Pause Job
      description: Pause a Job execution
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jobs"
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "No such job"
      tags:
        - Jobs
  /api/v1/jobs/{uuid}/resume/:
    post:
      operationId: Resume Job
      description: Resume Job execution
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jobs"
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "No such job"
      tags:
        - Jobs
  /api/v1/jobs/{uuid}/uptime/:
    get:
      operationId: Job uptime
      description: Returns a job uptime (Percentage)
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsUptime"
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "No such job"
      tags:
        - Jobs
  /api/v1/jobs/{uuid}/response/:
    get:
      operationId: Job response time
      description: Job response time (avg. seconds)
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "No such job"
      tags:
        - Jobs
  /api/v1/jobs/{uuid}/favicon/:
    get:
      operationId: favicon
      description: "Update favicon URL with the latest one."
      parameters:
        - name: uuid
          in: path
          required: true
          description: A unique value identifying this Jobs.
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jobs"
          description: ""
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "Job not found"
      tags:
        - Jobs
...
security:
  - bearerAuth: []
components:
  responses:
    UnauthorizedError:
      description: API token is missing or invalid
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      description: API uses bearer authentication mechanism
  schemas:
...
    Jobs:
      type: object
      properties:
        uuid:
          type: string
          readOnly: true
        url:
          type: string
          format: uri
          maxLength: 200
          pattern:
            "^(?:[a-z0-9.+-]*)://(?:[^\\s:@/]+(?::[^\\s:@/]*)?@)?(?:(?:0|25[0-5]|2[0-4]\\\
            d|1\\d?\\d?|[1-9]\\d?)(?:\\.(?:0|25[0-5]|2[0-4]\\d|1\\d?\\d?|[1-9]\\d?)){3}|\\\
            [[0-9a-f:.]+\\]|([a-z\xA1-\uFFFF0-9](?:[a-z\xA1-\uFFFF0-9-]{0,61}[a-z\xA1\
            -\uFFFF0-9])?(?:\\.(?!-)[a-z\xA1-\uFFFF0-9-]{1,63}(?<!-))*\\.(?!-)(?:[a-z\xA1\
            -\uFFFF-]{2,63}|xn--[a-z0-9]{1,59})(?<!-)\\.?|localhost))(?::\\d{2,5})?(?:[/?#][^\\\
            s]*)?\\z"
        title:
          type: string
          maxLength: 50
        state:
          type: boolean
        headers:
          type: object
        notifiers:
          type: array
          items:
            type: string
        verify_ssl:
          type: boolean
        healthy:
          type: boolean
        interval:
          type: integer
          maximum: 2147483647
          minimum: 0
        success_status:
          type: array
          items:
            type: integer
            maximum: 2147483647
        check_redirect:
          type: boolean
        favicon_url:
          type: string
      required:
        - url
        - title
    JobsUptime:
      type: object
      properties:
        uptime:
          type: string
          readOnly: true
      required:
        - uptime
    JobsResponse:
      type: object
      properties:
        response:
          type: string
          readOnly: true
      required:
        - response
...